requirements/requirements.txt
requests
python-dateutil
cryptography
urllib3
python-dotenv
websockets
datetime
selenium
beautifulsoup4
pytz
pandas
numpy
uuid
fake-useragent
python-time
random
fake_useragent
pytho-dotenv
webdriver-manager
scikit-learn

